// types/Auth.ts
export interface CreateAdminUserProps {
  /**
   * The email address of the new admin user.
   */
  readonly name: string

  /**
   * The email address of the new admin user.
   */
  readonly email: string

  /**
   * The name of the role of the new admin user.
   */
  readonly role: string
}